requirements.txt
dearpygui==1.10.1
numpy==1.24.2
opensimplex==0.4.5
Pillow